tools/requirements.txt
# Local Embedding Server & PDF OCR Tools Requirements
# Install with: pip install -r tools/requirements.txt

# Embedding Server
fastapi>=0.115.0
uvicorn>=0.32.0
sentence-transformers>=3.3.1
torch>=2.5.0

# PDF Processing & OCR
pytesseract>=0.3.10
pdf2image>=1.17.0
pdfplumber>=0.11.0
Pillow>=10.0.0

# Data Processing
pandas>=2.2.0
requests>=2.32.0

# Optional: Advanced table extraction (requires Java)
# camelot-py[cv]>=0.11.0
# tabula-py>=2.9.0
